i18n/requirements.txt
gitpython
langchain
openai
langchain-openai